+++ macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
 export "FLUTTER_ROOT=/home/root1/flutter"
-export "FLUTTER_APPLICATION_PATH=/home/root1/Documents/me/amazon_like_filter"
+export "FLUTTER_APPLICATION_PATH=/home/root1/Documents/me/amazon_like_filter-main"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"
 export "FLUTTER_BUILD_NAME=0.0.1"